feat: move tranche fill level to own file
